[WIP] app arguments edit & refacing
