+++ Ui/Model/Protocol/AppProtocol.cs
@@ -1,0 +1,223 @@
+﻿using System;
+using System.Diagnostics;
+using System.IO;
+using Newtonsoft.Json;
+using _1RM.Model.Protocol.Base;
+using _1RM.Utils;
+using Shawn.Utils;
+using Shawn.Utils.Wpf;
+using Shawn.Utils.Wpf.FileSystem;
+using _1RM.View;
+using System.Collections.ObjectModel;
+
+namespace _1RM.Model.Protocol
+{
+    public class LocalApp : ProtocolBase
+    {
+        public LocalApp() : base("APP", "APP.V1", "APP")
+        {
+        }
+
+        public override bool IsOnlyOneInstance()
+        {
+            return false;
+        }
+
+
+        private string _appSubTitle = "";
+        public string AppSubTitle
+        {
+            get => _appSubTitle;
+            set => SetAndNotifyIfChanged(ref _appSubTitle, value);
+        }
+
+        private string _exePath = "";
+        public string ExePath
+        {
+            get => _exePath;
+            set => SetAndNotifyIfChanged(ref _exePath, value);
+        }
+
+        private string _appProtocolDisplayName = "";
+        public string AppProtocolDisplayName
+        {
+            get => _appProtocolDisplayName;
+            set => SetAndNotifyIfChanged(ref _appProtocolDisplayName, value);
+        }
+
+        public override string GetProtocolDisplayName()
+        {
+            if (string.IsNullOrEmpty(_appProtocolDisplayName))
+                return base.GetProtocolDisplayName();
+            return _appProtocolDisplayName;
+        }
+
+        [Obsolete]
+        private string _arguments = "";
+        [Obsolete]
+        public string Arguments
+        {
+            get => _arguments;
+            set => SetAndNotifyIfChanged(ref _arguments, value);
+        }
+
+        private ObservableCollection<AppArgument> _argumentList = new ObservableCollection<AppArgument>();
+        public ObservableCollection<AppArgument> ArgumentList
+        {
+            get => _argumentList;
+            set => SetAndNotifyIfChanged(ref _argumentList, value);
+        }
+
+
+        private bool _runWithHosting = false;
+        public bool RunWithHosting
+        {
+            get => _runWithHosting;
+            set => SetAndNotifyIfChanged(ref _runWithHosting, value);
+        }
+        public string DemoArgumentsString => GetDemoArguments();
+
+        public override ProtocolBase? CreateFromJsonString(string jsonString)
+        {
+            try
+            {
+                var app = JsonConvert.DeserializeObject<LocalApp>(jsonString);
+                return app;
+            }
+            catch (Exception e)
+            {
+                SimpleLogHelper.Debug(e);
+                return null;
+            }
+        }
+
+
+        protected override string GetSubTitle()
+        {
+            return string.IsNullOrEmpty(AppSubTitle) ? $"{this.ExePath}" : AppSubTitle;
+        }
+
+        public override double GetListOrder()
+        {
+            return 100;
+        }
+
+        public string GetArguments()
+        {
+            return AppArgument.GetArgumentsString(ArgumentList, false);
+        }
+
+
+        public string GetDemoArguments()
+        {
+            return AppArgument.GetArgumentsString(ArgumentList, true);
+        }
+
+
+        private RelayCommand? _cmdSelectExePath;
+        [JsonIgnore]
+        public RelayCommand CmdSelectExePath
+        {
+            get
+            {
+                return _cmdSelectExePath ??= new RelayCommand((o) =>
+                {
+                    string initPath;
+                    try
+                    {
+                        initPath = new FileInfo(ExePath).DirectoryName!;
+                    }
+                    catch (Exception)
+                    {
+                        initPath = Environment.CurrentDirectory;
+                    }
+
+                    var path = SelectFileHelper.OpenFile(filter: "Exe|*.exe", initialDirectory: initPath, currentDirectoryForShowingRelativePath: Environment.CurrentDirectory);
+                    if (path == null) return;
+                    ExePath = path;
+                });
+            }
+        }
+
+        private RelayCommand? _cmdSelectArgumentFile;
+        [JsonIgnore]
+        public RelayCommand CmdSelectArgumentFile
+        {
+            get
+            {
+                return _cmdSelectArgumentFile ??= new RelayCommand((o) =>
+                {
+                    string initPath;
+                    try
+                    {
+                        initPath = new FileInfo(o?.ToString() ?? "").DirectoryName!;
+                    }
+                    catch (Exception)
+                    {
+                        initPath = Environment.CurrentDirectory;
+                    }
+                    var path = SelectFileHelper.OpenFile(initialDirectory: initPath, currentDirectoryForShowingRelativePath: Environment.CurrentDirectory);
+                    if (path == null) return;
+                    //Arguments = path;
+                });
+            }
+        }
+
+        private RelayCommand? _cmdPreview;
+        [JsonIgnore]
+        public RelayCommand CmdPreview
+        {
+            get
+            {
+                return _cmdPreview ??= new RelayCommand((o) =>
+                {
+                    MessageBoxHelper.Info(ExePath + " " + GetArguments(), ownerViewModel: IoC.Get<MainWindowViewModel>());
+                });
+            }
+        }
+
+        private RelayCommand? _cmdTest;
+        [JsonIgnore]
+        public RelayCommand CmdTest
+        {
+            get
+            {
+                return _cmdTest ??= new RelayCommand((o) =>
+                {
+                    try
+                    {
+                        Process.Start(ExePath, AppArgument.GetArgumentsString(ArgumentList, false));
+                        //    StartInfo =
+                        //{
+                        //    FileName = "cmd.exe",
+                        //    UseShellExecute = false,
+                        //    RedirectStandardInput = true,
+                        //    RedirectStandardOutput = true,
+                        //    RedirectStandardError = true,
+                        //    CreateNoWindow = true
+                        //}
+                        //var p = new Process
+                        //{
+                        //    StartInfo =
+                        //{
+                        //    FileName = "cmd.exe",
+                        //    UseShellExecute = false,
+                        //    RedirectStandardInput = true,
+                        //    RedirectStandardOutput = true,
+                        //    RedirectStandardError = true,
+                        //    CreateNoWindow = true
+                        //}
+                        //};
+                        //p.Start();
+                        //p.StandardInput.WriteLine(GetCmd());
+                        //p.StandardInput.WriteLine("exit");
+                    }
+                    catch (Exception e)
+                    {
+                        MessageBoxHelper.ErrorAlert(e.Message);
+                    }
+                });
+            }
+        }
+    }
+}
