+++ Ui/View/Editor/Forms/AppForm.xaml
@@ -29,7 +29,7 @@
             <StackPanel Orientation="Horizontal">
                 <TextBlock Style="{StaticResource EditorRowGridTitle}" Text="{DynamicResource 'Cmd parameter'}"></TextBlock>
                 <Grid Style="{StaticResource EditorRowGridInput}">
-                    <TextBox Text="{Binding ArgumentList, UpdateSourceTrigger=PropertyChanged}"></TextBox>
+                    <TextBox Text=""></TextBox>
                 </Grid>
 
                 <Button Style="{StaticResource ButtonPrimaryStyle}"
